requirements.txt
pyzipper
pycryptodome